+++ src/SPL/MatrixBalikan.java
@@ -8,14 +8,14 @@ public class MatrixBalikan {
         // Fungsi mencari solusi SPL dengan metode MatrixBalikan
         // Prekondisi matrixkoef dan matrixres tak kosong
         // KAMUS LOKAL
-        float[] solusi; //
+        double[] solusi; //
         String[] solusi_string; //
         Matrix tesmatriks = new Matrix(matrixkoef);
         int i; //indeks
         //ALGORITMA
         if (matrixkoef.getNCol()>=2){
             if (Matrix.isInvertible(tesmatriks)){ // Jika punya solusi unik
-                solusi = new float[matrixkoef.getNCol()];
+                solusi = new double[matrixkoef.getNCol()];
                 solusi_string = new String[matrixkoef.getNCol()];
                 Matrix inverseMatrix = Inverse.getInversebyOBE(matrixkoef);
                 Matrix matrixSolusi = Matrix.multiplyMatrix(inverseMatrix, matrixres);
@@ -23,13 +23,13 @@ public class MatrixBalikan {
                     solusi[i] = matrixSolusi.getElmtContent(i, 0);
                 }
             } else { // Jika tak punya matrix balikan
-                solusi = new float[0]; //Tak punya balikan
+                solusi = new double[0]; //Tak punya balikan
                 solusi_string = new String[0];
             }
 
         }
         else{ //matrixkoef.getNCol()==1
-            solusi = new float[matrixkoef.getNCol()];
+            solusi = new double[matrixkoef.getNCol()];
             solusi_string = new String[matrixkoef.getNCol()];
             if ((matrixkoef.getElmtContent(0, 0) == 0) && (matrixres.getElmtContent(0, 0)!=0)){
                 solusi =  GaussJordan.getlistUndef(1); //Jika tak punya solusi
@@ -41,7 +41,7 @@ public class MatrixBalikan {
         }
 
         for (i=0;i<solusi.length;i++){
-            solusi_string[i] = String.valueOf(solusi[i]);
+            solusi_string[i] = String.valueOf((double)solusi[i]);
         } 
         return solusi_string;
     }
